extensively test urls

+++ tasks.py
@@ -140,7 +140,7 @@ def docker_clean(ctx, no_sudo=False, all=False, volumes=False, networks=False, i
 # Translations
 #
 @task
-def i18n(ctx, compile=False, edit=False, lang='pt_BR'):
+def i18n(ctx, compile=False, edit=False, lang='pt_BR', keep_pot=False):
     """
     Extract messages for translation.
     """
@@ -162,8 +162,9 @@ def i18n(ctx, compile=False, edit=False, lang='pt_BR'):
         print(f'Update locale {lang} with Jinja2 messages')
         ctx.run(f'msgmerge locale/{lang}/LC_MESSAGES/django.po locale/join.pot -U')
 
-        print('Cleaning up')
-        ctx.run('rm locale/*.pot')
+        if not keep_pot:
+            print('Cleaning up')
+            ctx.run('rm locale/*.pot')
 
 
 #
